Modularize build
Move char classes related code from jflex.core to jflex.core.unicode
Move Out from jflex.core to jflex.logging
