+++ java/jflex/testing/testsuite/BUILD
@@ -13,6 +13,7 @@ java_library(
         "//java/jflex/util/javac",
         "//jflex",
         "//jflex/src/main/java/jflex/generator",
+        "//jflex/src/main/java/jflex/logging",
         "//third_party/com/google/guava",
         "//third_party/com/google/truth",
         "//third_party/junit",
