+++ jflex/src/main/java/jflex/BUILD
@@ -8,9 +8,12 @@ java_library(
         "//jflex/src/main/java/jflex/anttask",
         "//jflex/src/main/java/jflex/base",
         "//jflex/src/main/java/jflex/core",
+        "//jflex/src/main/java/jflex/core/unicode",
+        "//jflex/src/main/java/jflex/env",
         "//jflex/src/main/java/jflex/exceptions",
         "//jflex/src/main/java/jflex/generator",
         "//jflex/src/main/java/jflex/gui",
         "//jflex/src/main/java/jflex/l10n",
+        "//jflex/src/main/java/jflex/logging",
     ],
 )
