+++ jflex/src/main/java/jflex/generator/BUILD
@@ -5,10 +5,12 @@ java_library(
     deps = [
         "//jflex/src/main/java/jflex/base",
         "//jflex/src/main/java/jflex/core",
+        "//jflex/src/main/java/jflex/core/unicode",
         "//jflex/src/main/java/jflex/env",
         "//jflex/src/main/java/jflex/exceptions",
         "//jflex/src/main/java/jflex/io",
         "//jflex/src/main/java/jflex/l10n",
+        "//jflex/src/main/java/jflex/logging",
         "//jflex/src/main/java/jflex/performance",
         "//jflex/src/main/java/jflex/scanner",
     ],
