+++ jflex/src/main/java/jflex/gui/BUILD
@@ -10,5 +10,6 @@ java_library(
         "//jflex/src/main/java/jflex/exceptions",
         "//jflex/src/main/java/jflex/generator",
         "//jflex/src/main/java/jflex/l10n",
+        "//jflex/src/main/java/jflex/logging",
     ],
 )
