+++ jflex/src/main/java/jflex/state/BUILD
@@ -1,0 +1,12 @@
+java_library(
+    name = "state",
+    srcs = [
+        "StateSet.java",
+        "StateSetEnumerator.java",
+    ],
+    visibility = ["//jflex:__subpackages__"],
+    deps = [
+        "//jflex/src/main/java/jflex/env",
+        "//jflex/src/main/java/jflex/logging",
+    ],
+)
