+++ jflex/src/test/java/jflex/core/BUILD
@@ -1,43 +1,10 @@
 java_test(
-    name = "CharClassesTest",
-    srcs = ["CharClassesTest.java"],
-    deps = [
-        "//jflex/src/main/java/jflex/chars",
-        "//jflex/src/main/java/jflex/core",
-        "//third_party/com/google/truth",
-    ],
-)
-
-java_test(
-    name = "IntCharSetQuickcheck",
-    srcs = [
-        "IntCharSetGen.java",
-        "IntCharSetQuickcheck.java",
-    ],
-    deps = [
-        "//jflex/src/main/java/jflex/chars",
-        "//jflex/src/main/java/jflex/core",
-        "//third_party/com/google/truth",
-        "//third_party/com/pholser/quickcheck",
-    ],
-)
-
-java_test(
-    name = "IntCharSetTest",
-    srcs = ["IntCharSetTest.java"],
-    deps = [
-        "//jflex/src/main/java/jflex/chars",
-        "//jflex/src/main/java/jflex/core",
-        "//third_party/com/google/truth",
-    ],
-)
-
-java_test(
     name = "RegExpTests",
     srcs = ["RegExpTests.java"],
     deps = [
         "//jflex/src/main/java/jflex/chars",
         "//jflex/src/main/java/jflex/core",
+        "//jflex/src/main/java/jflex/core/unicode",
         "//third_party/com/google/truth",
     ],
 )
