+++ jflex/src/test/java/jflex/core/unicode/BUILD
@@ -1,0 +1,36 @@
+java_test(
+    name = "CharClassesTest",
+    srcs = ["CharClassesTest.java"],
+    deps = [
+        "//jflex/src/main/java/jflex/chars",
+        "//jflex/src/main/java/jflex/core",
+        "//jflex/src/main/java/jflex/core/unicode",
+        "//third_party/com/google/truth",
+    ],
+)
+
+java_test(
+    name = "IntCharSetQuickcheck",
+    srcs = [
+        "IntCharSetGen.java",
+        "IntCharSetQuickcheck.java",
+    ],
+    deps = [
+        "//jflex/src/main/java/jflex/chars",
+        "//jflex/src/main/java/jflex/core",
+        "//jflex/src/main/java/jflex/core/unicode",
+        "//third_party/com/google/truth",
+        "//third_party/com/pholser/quickcheck",
+    ],
+)
+
+java_test(
+    name = "IntCharSetTest",
+    srcs = ["IntCharSetTest.java"],
+    deps = [
+        "//jflex/src/main/java/jflex/chars",
+        "//jflex/src/main/java/jflex/core",
+        "//jflex/src/main/java/jflex/core/unicode",
+        "//third_party/com/google/truth",
+    ],
+)
