+++ jflex/src/test/java/jflex/core/unicode/IntCharSetQuickcheck.java
@@ -7,7 +7,7 @@
  *                                                                         *
  * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * */
 
-package jflex.core;
+package jflex.core.unicode;
 
 import static com.google.common.truth.Truth.assertThat;
