+++ jflex/src/test/java/jflex/core/unicode/IntCharSetTest.java
@@ -1,4 +1,4 @@
-package jflex.core;
+package jflex.core.unicode;
 
 import static com.google.common.truth.Truth.assertThat;
 import static com.google.common.truth.Truth.assertWithMessage;
